Clean up
remove duped code and make things happy
STATELESS
